+++ src/java/controller/DeleteUser.java
@@ -7,7 +7,8 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import model.dbManager.Manager;
+import model.Manager;
+import model.dataTypes.User;
 
 @WebServlet("/DeleteUser")
 public class DeleteUser extends HttpServlet {
@@ -25,7 +26,14 @@ public class DeleteUser extends HttpServlet {
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		try {
 			int userID = Integer.parseInt(request.getParameter("userID"));
-			Manager.deleteUser(userID);
+			String email = request.getParameter("email");
+			String password = request.getParameter("password");
+			User user = Manager.getUser(email, password);
+			
+			if(user.getType().equals("admin"))
+				Manager.deleteUser(userID);
+			else
+				throw new RuntimeException("Trying to hack me, eh?");
 		}
 		catch(RuntimeException e) {
 			response.getWriter().write(e.getMessage());
